Resolved cast warnings (safe explicit casting)

+++ System/private/vital.cpp
@@ -23,10 +23,10 @@
 
 namespace Vital {
     unsigned int getSystemTick() {
-        return std::chrono::high_resolution_clock::now().time_since_epoch().count()/1000000;
+        return static_cast<unsigned int>(std::chrono::high_resolution_clock::now().time_since_epoch().count() / 1000000);
     }
     unsigned int getApplicationTick() {
         vTick = vTick || getSystemTick();
-        return getSystemTick() - vTick;
+        return static_cast<unsigned int>(getSystemTick() - vTick);
     }
 }
